Fix chord simulator
Added 2nd hash function to distribute replicas of segments
